Refactor credentials module utilities
